backend/requirements.txt
# Web Framework
flask==2.0.1
werkzeug==2.0.1
gunicorn==20.1.0

# Database
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.41

# Security
PyJWT==2.8.0
python-dotenv==0.19.0

# CORS
flask-cors==3.0.10

# Dependencies
click==8.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1

# Production Database (uncomment for PostgreSQL)
# psycopg2-binary==2.9.1
